clients/python_publisher/requirements.txt
paho-mqtt==2.1.0
PyYAML==6.0.2
cryptography==43.0.1